requirements.txt
selenium
requests-async
beautifulsoup4
lxml
chromedriver-binary-auto